Add a skeleton script to download FEC data
Add get_fec_data.py, which contains a commented plan for identifying
all of the bulk data download files from the FEC website. It turned out
to be difficult to identify each of the data types just from the html
content of the page, so I later changed my approach.